Adding FileReadSecretsFetcher to support Vault Agent Injector usage moved secrets client logic from root.go to secrets pkg left off at K8sConfig Process Secrets Injections which currently accepts PVC Client
